all: modify the program to be interpreted by the toplevel
The reason for this, is that trying to link with chipmunk (for reasons I
don't understand) causes ocamlc to attempt compilation of  a custom
runtime, which attempts to statically link with some ocaml-provided
libraries, which fails because relocation fails. Recompiling the
requisite OCaml libraries to be position independent would avoid the need
to do relocation and thefore avoid this issue, but would also require
the TA to do the same, which is not reasonable. For whatever reason,
linking with OpenGL (which also caused ocamlc to build a custom runtime)
did not cause this issue. After many hours of debugging, the most viable
option seems to be to interpret the code with the ocaml toplevel
instead, because the ocaml toplevel seems to have no problem loading the
chipmunk bytecode library and calling chipmunk functions. Go figure.